requirements/base.txt
Django==1.6.5
South==0.8.4
Unipath==1.0
boto==2.28.0
django-bootstrap3==4.6.0
django-crispy-forms==1.4.0
django-extra-views==0.6.4
django-markdown==0.6.0
django-markdown-deux==1.0.4
django-model-utils==2.0.3
django-vanilla-views==1.0.3
markdown2
psycopg2==2.5.3
six==1.6.1
sqlparse==0.1.11
django.js==0.8.1
django-eztables==0.3.2
